+++ index.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
+import ChunkErrorBoundary from './components/ChunkErrorBoundary';
 
 
 const rootElement = document.getElementById('root');
@@ -12,6 +13,8 @@ if (!rootElement) {
 const root = ReactDOM.createRoot(rootElement);
 root.render(
   <React.StrictMode>
-    <App />
+    <ChunkErrorBoundary>
+      <App />
+    </ChunkErrorBoundary>
   </React.StrictMode>
 );